+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/compose/ListButton.kt
@@ -14,6 +14,7 @@ import androidx.compose.material3.CardDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
@@ -34,6 +35,7 @@ fun ListButton(
     Row(
         modifier.hoverable(interactionSource,!selected),
         horizontalArrangement = Arrangement.spacedBy(8.dp),
+        verticalAlignment = Alignment.CenterVertically
     ) {
         leadingContent?.invoke()
         var shape = CardDefaults.outlinedShape